Region segmentation

+++ lib/walet.h
@@ -178,7 +178,7 @@ struct Region{
 
 struct Row {
 	uint16		x;			//The start row on axis X
-	uint16 		y;			//The start point
+	uint32 		y;			//The start point
 	uchar 		c[4];		//The colors
 	uint16 		length;		//The length of row
 	Region		*reg;		//Pointer to the region
